pyproject.toml
[project]
name = "listr"
version = "0.0.0"
authors = [{name = "James Keal", email = "james@keal.au"}]

dependencies = [
  "discord.py",
  "pydantic",
  "pyyaml",
]

[project.optional-dependencies]
dev = [
  "black",
  "isort",
  "mypy",
  "pip-tools",
  "pytest",
  "types-pyyaml",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

